feat: add response generation function
